+++ model_navigator/api/pytriton.py
@@ -95,12 +95,14 @@ class ModelConfig:
         max_batch_size: The maximal batch size that would be handled by model.
         batcher: Configuration of Dynamic Batching for the model.
         response_cache: Flag to enable/disable response cache for the model
+        decoupled: Flag to enable/disable decoupled transaction policy
     """
 
     batching: bool = True
     max_batch_size: int = 4
     batcher: DynamicBatcher = dataclasses.field(default_factory=DynamicBatcher)
     response_cache: bool = False
+    decoupled: bool = False
 
 
 @dataclasses.dataclass(frozen=True)